split `alert` module in two
